agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Pablo", email = "carebymedia@gmail.com" }
]
dependencies = [
    "openai>=1.56.2",
    "web3>=7.6.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
